fix(vc2): mscn_2048d bound path -- adaptive box-remainder zono in input-split
The multi-sub route's dense per-batch zono needs 130+ GiB on
mscn_2048d (mul/reciprocal diagonal columns; the code's own comment)
and OOMed every batch into the slow chunked-CROWN degrade -- the row's
official budget is 20s. input_split_bab now switches PERMANENTLY to
box_remainder='all' after the first dense OOM: mul/reciprocal collapse
into the rad vector, the generator count stays ~n_wide.

Also fixes a rad-mode bug the probe found: concat with an empty-sym
part (a mul output under 'all') built a float index tensor and crashed
the advanced indexing.

Measured locally (CPU, 8G cap): 16 real cardinality_0_500_2048
sub-boxes bound in 1.6s with 8/16 refuted at ROOT by crown over the
remainder-zono planes. Official row is itself a tool DISAGREEMENT
(v1 sat@80.8s -- the within-tol-CE pattern -- vs ab unsat@11.6s);
under the strict policy vc2's target is ab's side.

+++ src/vibecheck2/core/forward.py
@@ -537,7 +537,13 @@ def zono(net, lo, hi, return_state=False, record=None, clamp_bounds=None,
                                      op.params['positions']):
                 p = torch.as_tensor(pos, device=dev)
                 c2[:, p] = zp.c
-                G2[:, p.unsqueeze(1), torch.as_tensor(cols, device=dev)] = zp.G
+                # dtype=long explicitly: an EMPTY cols list (a part whose
+                # symbols all live in the remainder, e.g. a mul output
+                # under box_remainder='all') otherwise infers float and
+                # crashes the advanced indexing
+                G2[:, p.unsqueeze(1),
+                   torch.as_tensor(cols, device=dev,
+                                   dtype=torch.long)] = zp.G
                 if rad2 is not None:
                     rad2[:, p] = zp.rad
             state[name] = ZonoState(c2, G2, syms, rad2)
